Комментарии теперь находятся внутри PostDTO. Добавлена возможность получения поста по id, добавлено несколько методов для упрощения кода

+++ src/main/java/org/app/breeze/controller/CommentController.java
@@ -25,7 +25,7 @@ public class CommentController {
     private PostRepository postRepository;
 
     @GetMapping()
-    public ResponseEntity<List<CommentDTO>> getCommentsForPost(@PathVariable int postId) {
+    public ResponseEntity<List<CommentDTO>> getCommentsForPost(@PathVariable Long postId) {
         List<CommentDTO> comments = commentService.getCommentsForPost(postId);
         return ResponseEntity.ok(comments);
     }
